Lab_9/Lab_9/requirements.txt
Flask
gunicorn
sqlalchemy
psycopg2-binary
alembic
marshmallow
flask-bcrypt
Flask-JWT
Flask-Testing
coverage